+++ example_impl/example_testsuite/example_testsuite.go
@@ -12,10 +12,15 @@ import (
 // TODO example of parameterizing your image
 type ExampleTestsuite struct {}
 
-func (e ExampleTestsuite) GetTests() map[string]testsuite.Test {
+func (suite ExampleTestsuite) GetTests() map[string]testsuite.Test {
 	return map[string]testsuite.Test{
 		"singleNodeExampleTest": SingleNodeExampleTest{},
 		"fixedSizeExampleTest": FixedSizeExampleTest{},
 	}
 }
+
+func (suite ExampleTestsuite) GetNetworkWidthBits() uint32 {
+	return 8
+}
+