+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Added another test to demonstrate an initial network setup
 * Adding copyright headers
 * Renamed tests to have more descriptive names
+* When asked about test suite data, send back a JSON of test suite metadata (rather than just a list of test names)